requirements.txt
dash-bootstrap-components
dash-daq
gunicorn
librosa
pandas
scikit-learn
scikit-maad
tensorflow